+++ CHANGES.md
@@ -6,6 +6,7 @@
 
 - `CesiumRasterOverlays::WebMapServiceRasterOverlay` will no longer attempt to add a default `styles` parameter to the URL if the user has already specified one.
 - `SharedAssetDepot` will no longer cache asset loads that fail with an exception / `Future` rejection, allowing them to be retried. Other types of load failures are cached as before.
+- Fixed a bug in `JsonHelpers::GetInt64OrDefault` and `GetUint64OrDefault` that made them return the default for a value larger than the maximum value of a 32-bit integer.
 
 ##### Additions :tada: